requirements.txt
djangorestframework
django-cors-headers
pillow
pytils
django-ckeditor
